CIT-30 Add Arbre A champs

+++ src/main/java/com/citronix/citronix/repository/ArbreRepository.java
@@ -3,5 +3,5 @@ package com.citronix.citronix.repository;
 import com.citronix.citronix.entity.Arbre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArbreRepository extends JpaRepository<Arbre, Integer> {
+public interface ArbreRepository extends JpaRepository<Arbre, Long> {
 }
